优化价格推送逻辑和图表策略，重构价格获取逻辑优先使用K线缓存提升性能，优化Pivot Order Blocks绘制代码减少88%填充线提升渲染效率，更新实时行情间隔支持小数配置。所有功能已测试通过，代码质量良好。

+++ stock_app_service/app/core/config.py
@@ -83,7 +83,7 @@ LOG_LEVEL = os.getenv("LOG_LEVEL", "INFO")
 MAX_HISTORY_RECORDS = int(os.getenv("MAX_HISTORY_RECORDS", "1000"))
 
 # 实时行情配置（仅Tushare）
-REALTIME_UPDATE_INTERVAL = int(os.getenv("REALTIME_UPDATE_INTERVAL", "1"))  # 实时更新周期，单位：分钟，默认1分钟
+REALTIME_UPDATE_INTERVAL = float(os.getenv("REALTIME_UPDATE_INTERVAL", "1"))  # 实时更新周期，单位：分钟，默认1分钟，支持小数如0.5表示30秒
 
 
 class Settings(BaseModel):
@@ -117,7 +117,7 @@ class Settings(BaseModel):
     DEFAULT_AI_MODEL: str = DEFAULT_AI_MODEL
     
     # 实时行情配置（仅Tushare）
-    REALTIME_UPDATE_INTERVAL: int = REALTIME_UPDATE_INTERVAL
+    REALTIME_UPDATE_INTERVAL: float = REALTIME_UPDATE_INTERVAL
     
     class Config:
         case_sensitive = True
